pyproject.toml
[tool.poetry]
name = "django-steamboat"
version = "0.1.0"
description = "A custom sequence information management system powered by Django"
authors = ["Robert A. Petit III <robbie.petit@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "==4.2.*"
psycopg = ">=3.1.8"
django-environ = "^0.11.2"
python-slugify = "^8.0.4"
pillow = "^10.3.0"
whitenoise = "^6.6.0"
django-model-utils = "^4.5.0"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2024.2"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.27.2"
django-redis = "^5.4.0"
rcssmin = "^1.1.2"
argon2-cffi = "^23.1.0"
django-allauth = {extras = ["mfa"], version = "^0.61.1"}
uvicorn = {extras = ["standard"], version = "^0.29.0"}
hiredis = "^2.3.2"
django-unfold = "^0.22.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
django-debug-toolbar = "^4.3.0"
django-extensions = "^3.2.3"
django-coverage-plugin = "^3.1.0"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"

[tool.black]
line-length = 79
force-exclude = '''
    (
        config
        | migrations
    )
    '''

[tool.isort]
profile = "black"
skip_glob = "**/migrations"
extend_skip_glob = [".trash", "config"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

